Refactored SampleTests.cs. and R# templates. Test Method should not commit the Transaction...
Why? These are supposed to be Unit Tests, the database should be clean for each test method.
The helper method may commit the Transaction so the object is in the database to test but the Test Method should RollBack the transaction so the database is clean for the next test.
The Test Method wraps the helper method transaction so it will RollBack any committed transactions that happen inside it.